add mouse actions to controls modal
